llboursa_scraper/requirements.txt
browser-use
playwright
langchain-openai
python-dotenv
supabase
anthropic
